requirements.txt
amqp==5.1.1
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.5.2
async-timeout==4.0.2
cachetools==4.2.2
celery==5.2.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cssselect==1.1.0
DateTime==4.4
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.1
eventlet==0.33.1
fake-useragent==0.1.11
Flask==2.2.2
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
gspread==5.5.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lxml==4.8.0
MarkupSafe==2.1.1
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.4
parse==1.19.0
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.31
protobuf==4.21.5
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.3.4
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
self==2020.12.3
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.39
sqlparse==0.4.2
telegram-send==0.25
tenacity==8.0.1
tornado==6.1
tqdm==4.63.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.2
Werkzeug==2.2.2
wrapt==1.14.1
xmltodict==0.13.0
zipp==3.7.0
zope.interface==5.4.0
